+++ R/print_igraph.R
@@ -1,17 +1,16 @@
 #' @title Print graphs to terminal
 #' @description Prints an igraph object to terminal (different than the standard igraph method)
-#' @param x An igraph object
+#' @param object An igraph object
 #' @param ... additional arguments to print (ignored)
 #' @author David Schoch
 #' @export
-
-print.igraph <- function(x,...){
+str.igraph <- function(object,...){
   maxpr  <- getOption("width")
-  gattrs <- igraph::graph.attributes(x)
-  vattrs <- igraph::vertex.attributes(x)
-  eattrs <- igraph::edge.attributes(x)
-  comps  <- igraph::components(x)
-  dens   <- igraph::graph.density(x)
+  gattrs <- igraph::graph.attributes(object)
+  vattrs <- igraph::vertex.attributes(object)
+  eattrs <- igraph::edge.attributes(object)
+  comps  <- igraph::components(object)
+  dens   <- igraph::graph.density(object)
 
   #header
   if(!"name" %in% names(gattrs)){
@@ -21,10 +20,10 @@ print.igraph <- function(x,...){
   }
 
   head <- paste(toupper(gname)," ",
-                c("(undirected","(directed")[igraph::is.directed(x)+1],", ",
-                c("unweighted","weighted")[igraph::is.weighted(x)+1],", ",
+                c("(undirected","(directed")[igraph::is.directed(object)+1],", ",
+                c("unweighted","weighted")[igraph::is.weighted(object)+1],", ",
                 c("","signed, ")["sign"%in%names(eattrs)+1],
-                c("one-mode","two-mode")[igraph::is.bipartite(x)+1]," ",
+                c("one-mode","two-mode")[igraph::is.bipartite(object)+1]," ",
                 "network)",sep=""
   )
 
@@ -34,11 +33,11 @@ print.igraph <- function(x,...){
   short_delim <- "---\n"
 
   # graph details
-  n   <- paste0("Nodes: ",igraph::vcount(x))
-  m   <- paste0("Edges: ",igraph::ecount(x))
+  n   <- paste0("Nodes: ",igraph::vcount(object))
+  m   <- paste0("Edges: ",igraph::ecount(object))
   d   <- paste0("Density: ",ifelse(dens<1e-4,"<1e-4",round(dens,4)))
   cc  <- paste0("Components: ",comps$no)
-  iso <- paste0("Isolates: ",sum(igraph::degree(x)==0))
+  iso <- paste0("Isolates: ",sum(igraph::degree(object)==0))
   gstats <- paste(n,m,d,cc,iso,sep=", ")
   gstats <- paste0(strwrap(gstats),collapse = "\n")
   gstats <- paste0(gstats,"\n")
@@ -80,11 +79,11 @@ print.igraph <- function(x,...){
   if(length(eattrs)>0){
     eattr_str <- paste0(eattr_str,short_delim)
   }
-  if(igraph::ecount(x)>0){
-    edges <- igraph::get.edgelist(x)[1:min(c(10,igraph::ecount(x))),]
-    edges <- strwrap(paste0(apply(edges,1,paste0,collapse=c("->","--")[igraph::is.directed(x)+1]),collapse=" "))
+  if(igraph::ecount(object)>0){
+    edges <- igraph::get.edgelist(object)[1:min(c(10,igraph::ecount(object))),]
+    edges <- strwrap(paste0(apply(edges,1,paste0,collapse=c("->","--")[igraph::is.directed(object)+1]),collapse=" "))
     edges <- paste(edges,collapse = "\n")
-    if(igraph::ecount(x)>10){
+    if(igraph::ecount(object)>10){
       edges <- c("-Edges (first 10): \n ",edges)
     } else{
       edges <- c("-Edges: \n ",edges)