+++ NEWS.md
@@ -1,4 +1,11 @@
-# igraphUtils 0.1.0.9000
+# netUtils 0.5.0.9000
+
+* renamed package to `netUtils`
+* added `bipartite_from_data_frame()`
+* added `graph_from_multi_edgelist()` and `as_multi_adj()`
+* added `structural_equivalence()`
+
+# igraphUtils 0.1.0
 
 * Added `as_adj_list1()` and `as_adj_weighted()`
 * Added `clique_vertex_mat()`